api-django-drf/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
dnspython==2.7.0
mongoengine==0.29.1
pymongo==4.14.1
sqlparse==0.5.3
